Break up `util` into `io` and `hex`
